requirements.txt
asgiref==3.4.1
Django==3.2.5
django-crispy-forms==1.12.0
django-fontawesome-5==1.0.18
django-froala-editor==3.2.2
django-richtextfield==1.6.1
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djongo==1.3.6
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.3.1
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
whitenoise==5.3.0
